new convenience methods to advice and resize files, refactor tests
