+++ src/runtime/types.rs
@@ -53,6 +53,45 @@ pub enum ChunkType {
     V2 = 2,
 }
 
+#[derive(Clone, Copy, Debug)]
+pub enum Advice {
+    Normal = 0,
+    Sequential = 1,
+    Random = 2,
+    WillNeed = 3,
+    DontNeed = 4,
+    NoReuse = 5,
+}
+
+impl TryFrom<u8> for Advice {
+    type Error = crate::error::Error;
+
+    fn try_from(value: u8) -> Result<Self, crate::error::Error> {
+        match value {
+            0 => Ok(Advice::Normal),
+            1 => Ok(Advice::Sequential),
+            2 => Ok(Advice::Random),
+            3 => Ok(Advice::WillNeed),
+            4 => Ok(Advice::DontNeed),
+            5 => Ok(Advice::NoReuse),
+            _ => Err(crate::error::Error::InvalidArgument),
+        }
+    }
+}
+
+impl From<Advice> for u8 {
+    fn from(val: Advice) -> Self {
+        match val {
+            Advice::Normal => 0,
+            Advice::Sequential => 1,
+            Advice::Random => 2,
+            Advice::WillNeed => 3,
+            Advice::DontNeed => 4,
+            Advice::NoReuse => 5,
+        }
+    }
+}
+
 bitflags! {
     #[derive(Copy, Clone, Debug, PartialEq)]
     pub struct FdFlags: u16 {
